requirements.txt
Flask
mysql-connector-python
pytest
python-dotenv
pyjwt
bcrypt